requirements.txt
pydantic
google-generativeai
pydantic-settings